requirements.txt
google-cloud-api
python-telegram-bot
yargy
natasha
